+++ src/morph_kgc/__init__.py
@@ -18,7 +18,7 @@ from itertools import repeat
 
 from .args_parser import load_config_from_command_line
 from .mapping.mapping_parser import retrieve_mappings
-from .data_source.relational_database import setup_oracle
+from .data_source.relational_db import setup_oracle
 from .materializer import _materialize_mapping_group_to_set
 from .args_parser import load_config_from_argument
 from .constants import RML_TRIPLES_MAP_CLASS
@@ -96,7 +96,7 @@ def materialize_kafka(config, python_source=None):
         if not output_kafka_server or not output_kafka_topic:
             logging.error('Output Kafka server or topic is empty.')
             sys.exit()
-        
+
         kafka_producer = KafkaProducer(bootstrap_servers=output_kafka_server)
 
         if triples:
