+++ src/morph_kgc/__main__.py
@@ -16,7 +16,7 @@ from itertools import repeat
 from .args_parser import load_config_from_command_line
 from .materializer import _materialize_mapping_group_to_file
 from .materializer import _materialize_mapping_group_to_kafka
-from .data_source.relational_database import setup_oracle
+from .data_source.relational_db import setup_oracle
 from .utils import get_delta_time
 from .mapping.mapping_parser import retrieve_mappings
 from .constants import RML_TRIPLES_MAP_CLASS
